fix consistency of wavelength data type between files

+++ mkidpipeline/hdf/binlib.h
@@ -1,3 +1,3 @@
 long cparsebin(const char *fName, unsigned long max_len,
-int* baseline, int* wavelength, unsigned long* timestamp,
+int* baseline, float* wavelength, unsigned long* timestamp,
 unsigned int* ycoord, unsigned int* xcoord, unsigned int* roach);
